perf: simplify DashboardAppProvider
