+++ src/components/dashboard/DashboardAppProvider.tsx
@@ -1,5 +1,4 @@
 // Providers
-import { AppProvider } from '@toolpad/core/nextjs';
 import { AppRouterCacheProvider } from '@mui/material-nextjs/v15-appRouter';
 
 // Hooks
@@ -7,16 +6,14 @@ import {setRequestLocale} from 'next-intl/server';
 import {useTranslations} from 'next-intl';
 
 // components
-import { Suspense } from 'react';
 import { DashboardLayout } from '@toolpad/core/DashboardLayout';
+import AppProviderCustom from "@/components/dashboard/AppProviderCustom";
 import ToolbarActions from "@/components/dashboard/ToolbarActions";
 
 type Props = {
     children: ReactNode,
     locale: "en" | "fr"
 }
-
-import NavigationMenu from "./MenuEntries";
 
 // Types
 import type { ReactNode } from "react";
@@ -27,31 +24,22 @@ export default function DashboardAppProvider({children, locale} : Props) {
     setRequestLocale(locale);
 
     // Fetch labels
-    const t = useTranslations('dashboard');
+    const t = useTranslations('dashboard.toolbar');
 
-    // Generate navigation
-    const NAVIGATION = NavigationMenu({
-        backlog: t("menuEntries.backlog"),
-        dlcsView: t("menuEntries.gamesTabs.dlc"),
-        gamesCategoy: t("menuEntries.gamesKey"),
-        gamesView: t("menuEntries.gamesTabs.grid"),
-        links: t("menuEntries.links"),
-        planned: t("menuEntries.planningKey"),
-        seriesView: t("menuEntries.gamesTabs.list"),
-        stats: t("menuEntries.stats"),
-        tests: t("menuEntries.testsKey")
-    });
+    const toolbarActionsProps = {
+        settingsLabel: t("settings"),
+        darkLabel: t("modes.dark"),
+        englishLabel: t("languages.en"),
+        frenchLabel: t("languages.fr"),
+        languageTitle: t("languages.title"),
+        lightLabel: t("modes.light"),
+        modeTitle: t("modes.title"),
+        systemLabel: t("modes.system")
+    }
 
     return (
         <AppRouterCacheProvider options={{ enableCssLayer: true }}>
-            <Suspense fallback={<></>}>
-                <AppProvider 
-                    navigation={NAVIGATION}
-                    branding={{
-                        title: "GamesPassionFR",
-                        logo: <>&nbsp;</>
-                    }}
-                >
+            <AppProviderCustom>
                 <DashboardLayout 
                     defaultSidebarCollapsed 
                     slots={{
@@ -59,22 +47,12 @@ export default function DashboardAppProvider({children, locale} : Props) {
                         toolbarActions: ToolbarActions
                     }}
                     slotProps={{
-                        toolbarActions: {
-                            settingsLabel: t("toolbar.settings"),
-                            darkLabel: t("toolbar.modes.dark"),
-                            englishLabel: t("toolbar.languages.en"),
-                            frenchLabel: t("toolbar.languages.fr"),
-                            languageTitle: t("toolbar.languages.title"),
-                            lightLabel: t("toolbar.modes.light"),
-                            modeTitle: t("toolbar.modes.title"),
-                            systemLabel: t("toolbar.modes.system")
-                        }
+                        toolbarActions: toolbarActionsProps
                     }}
                 >
                     {children}
                 </DashboardLayout>
-                </AppProvider>
-            </Suspense>
+            </AppProviderCustom>
         </AppRouterCacheProvider>
     );
 }
